fM - Added authenticator app support for 2FA
With the addition of email 2FA, this PR adds support for authenticator apps. Composer was installed to install apps via composer, but it should be decided if the fM project should require composer and users to install dependencies or if they deps would be included in this project. Therefore the PR does not include them nor the composer file.

This PR has the following dependencies:

* fM 6.0.0
* robthree/twofactorauth ^3.0 (via composer)
* bacon/bacon-qr-code ^2.0 (via composer)
* php-imagick